+++ g-fashion-domain/src/main/java/com/gfashion/domain/cart/GfCartTotals.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.json.JSONObject;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -14,16 +15,16 @@ import java.util.List;
 @NoArgsConstructor
 public class GfCartTotals {
 
-    private Integer grandTotal;
-    private Integer baseGrandTotal;
+    private BigDecimal grandTotal;
+    private BigDecimal baseGrandTotal;
     private Integer subtotal;
     private Integer baseSubtotal;
     private Integer discountAmount;
     private Integer baseDiscountAmount;
     private Integer subtotalWithDiscount;
     private Integer baseSubtotalWithDiscount;
-    private Integer shippingAmount;
-    private Integer baseShippingAmount;
+    private BigDecimal shippingAmount;
+    private BigDecimal baseShippingAmount;
     private Integer shippingDiscountAmount;
     private Integer baseShippingDiscountAmount;
     private Integer taxAmount;
@@ -32,8 +33,8 @@ public class GfCartTotals {
     private Integer shippingTaxAmount;
     private Integer baseShippingTaxAmount;
     private Integer subtotalInclTax;
-    private Integer shippingInclTax;
-    private Integer baseShippingInclTax;
+    private BigDecimal shippingInclTax;
+    private BigDecimal baseShippingInclTax;
     private String baseCurrencyCode;
     private String quoteCurrencyCode;
     private Integer itemsQty;
